requirements.txt
absl-py==2.2.2
antropy==0.1.9
asttokens==3.0.0
astunparse==1.6.3
certifi==2025.4.26
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
flatbuffers==25.2.10
fonttools==4.58.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==8.18.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jupyterlab_widgets==3.0.15
keras==3.10.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.43.0
lspopt==1.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mne==1.8.0
namex==0.0.9
numba==0.60.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
prompt_toolkit==3.0.51
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
pyriemann==0.8
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sleepecg==0.5.8
stack-data==0.6.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
tensorpac==0.6.5
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
yasa==0.6.5
zipp==3.21.0
